+++ preprocess/utils/__init__.py
@@ -1,2 +1,2 @@
 from preprocess.utils.helpers import trunc_helpers, maximal_weight_proportion, is_valid_solution
-from preprocess.utils.preprocess_transform import Preprocess, VectorMap
+from preprocess.utils.transforms import Preprocess, VectorMap
